requirements.txt
bioblend>=0.18.0
python-dotenv>=1.0.0 